minor prisma schema additions and rename file
